Add files via upload
This adds a button called "Add Random Hooks" in the Option panel that adds 30 hook on random places on the active object, as well as a Cloth sim. Play the Timeline and grab the hooks to create realistic folds in your cloth object. Then once you're done with that, apply the modifiers and remove hooks with the "Remove Hooks" button. Unfortunately both buttons are in separate tabs, sorry about that. Have fun!